iwlwifi: get rid of pointless casts

+++ drivers/net/wireless/iwlwifi/pcie/tx.c
@@ -1634,7 +1634,7 @@ void iwl_pcie_hcmd_complete(struct iwl_trans *trans,
 		struct page *p = rxb_steal_page(rxb);
 
 		meta->source->resp_pkt = pkt;
-		meta->source->_rx_page_addr = (unsigned long)page_address(p);
+		meta->source->_rx_page_addr = page_address(p);
 		meta->source->_rx_page_order = trans_pcie->rx_page_order;
 	}
